[sync] Enhancement: Refactoring Source Tree
- Move mutable and non-mutable buffers to collections module
- Added a few unit tests for non-mutable buffers
